refactor: move MRIRepository to repositories module
