+++ oxytcmri/tests/test_unit_domain.py
@@ -1,7 +1,7 @@
 from oxytcmri.domain.entities.subject import Subject, SubjectType
 from oxytcmri.domain.entities.center import Center
-from oxytcmri.domain.entities.mri import DTIMetric, Atlas, MRIExam, MRIRepository, MRIData
-from oxytcmri.domain.ports.repositories import SubjectRepository
+from oxytcmri.domain.entities.mri import DTIMetric, Atlas, MRIExam, MRIData
+from oxytcmri.domain.ports.repositories import SubjectRepository, MRIRepository
 from oxytcmri.domain.use_cases.compute_dti_normative_values import ComputeDTINormativeValues
 import pytest
 from typing import List, Optional
